NLOptControl.jl

NLOptControl.jl

The following link provides documentation all of the exported NLOptControl.jl functions:

NLOptControl.OCPdef! — Method.

n,r=OCPdef(mdl,n,s)

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 1/14/2017, Last Modified: 5/4/2017

source

NLOptControl.configure! — Method.

n = configure!(n::NLOpt,Ni=4,Nck=[3, 3, 7, 2];(:integrationMethod => :ps),(:integrationScheme => :lgrExplicit),(:finalTimeDV => false),(:tf => 1))

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 1/1/2017, Last Modified: 3/25/2017

source

NLOptControl.define! — Method.

n = define!(n,numStates=2,numControls=1,X0=[0.,1],XF=[0.,-1.],XL=[0.,-Inf],XU=[1/9,Inf],CL=[-Inf],CU=[Inf])

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 1/1/2017, Last Modified: 4/3/2017

Citations:

Initially Influenced by: S. Hughes. steven.p.hughes@nasa.gov Source: DecisionVector.m located here ––––––––––––––––––––––––––––––––––––––––––-

source

NLOptControl.defineSolver! — Method.

mdl=defineSolver!(n;name=:KNITRO,max_iter=1000,feastol_abs=1.0e-3,infeastol=1.0e-8,opttol_abs=1.0e-3)

TODO make an option to run solvers with default settings

figure out best correspondence between IPOPT and KNITRO settings

To debug KNITRO turn up the optput level

Try to tune KNITRO

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/9/2017, Last Modified: 4/3/2017

source

NLOptControl.defineTolerances! — Method.

defineTolerances!(n)

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/8/2017, Last Modified: 2/8/2017

source

NLOptControl.evalMaxDualInf — Method.

maxDualInf = evalMaxDualInf(n,r)

funtionality to evaluate maximum dual infeasibility of problem

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/13/2017, Last Modified: 2/13/2017

source

NLOptControl.integrate! — Method.

integrating JuMP variables

Expr=integrate!(mdl,n,u;(:mode=>:control)) Expr=integrate!(mdl,n,u,idx=1;C=0.5,(:variable=>:control),(:integrand=>:squared)) Expr=integrate!(mdl,n,r.u[:,1];D=rand(n.numStatePoints),(:variable=>:control),(:integrand=>:squared),(:integrandAlgebra=>:subtract)) #TODO fix D ::Array{JuMP.NonlinearParameter,1} –––––––––––––––––––––––––––––––––––––––––––

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 1/2/2017, Last Modified: 4/12/2017

source

NLOptControl.linearStateTolerances! — Method.

linearStateTolerances!(n)

the purpose of this function is to taper the tolerances on the constant state constraints

the idea is that when doing MPC, the final states are well within the bounds so that the next optimization is not initalized at an infeasible point

if you want a constant bond, set the slope to zero

default is a positive slope on the lower bound and a negative slope on the upper bound

this functionality in not needed for states like position, so you do not need to add a linearStateTol for all states

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/23/2017, Last Modified: 3/25/2017

source

NLOptControl.maxDF — Method.

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 12/8/2016, Last Modified: 3/29/2017

source

NLOptControl.minDF — Method.

maximum(x->maximum(x[:A]), dfs) -> consider

maximum(x->maximum(filter(y->y !== nothing, x[:A])), dfs)

minimum(x->maximum(filter(y->y !== nothing, x[:t])), dfs)

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 12/8/2016, Last Modified: 3/29/2017

source

NLOptControl.resultsDir! — Method.

description = string( " * ")

setupResults(r;results_name,description=description)

removes results folder and creates a new one

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/26/2017, Last Modified: 5/19/2017

source

NLOptControl.savePlantData — Method.

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/26/2017, Last Modified: 4/27/2017

source