VehicleModels.jl

VehicleModels.jl

The following link provides documentation all of the exported VehicleModels.jl functions:

VehicleModels.KinematicBicycle — Method.

dx=KinematicBicycle(n,x,u)

Original Authors: BARC Project, Berkely MPC Laboratory -> https://github.com/MPC-Berkeley/barc Modified for NLOptControl.jl by: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/9/2017, Last Modified: 5/31/2017

this vehicle model is controlled using steering angle and longitudinal acceleration

source

VehicleModels.ThreeDOFv1 — Method.

dx = ThreeDOFv1(n,x,u)

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 10/20/2017, Last Modified: 5/31/2017

this vehicle model is controlled using steering angle and speed

source

VehicleModels.ThreeDOFv1 — Method.

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 10/01/2016, Last Modified: 4/4/2017

source

VehicleModels.ThreeDOFv2 — Method.

dx = ThreeDOFv2(n,x,u)

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 10/20/2017, Last Modified: 5/31/2017

this vehicle model is controlled using steering rate and longitudinal jerk

source

VehicleModels.ThreeDOFv2 — Method.

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 10/01/2016, Last Modified: 4/11/2017

source