PrettyPlots.jl

PrettyPlots.jl

The following link provides documentation all of the exported PrettyPlots.jl functions:

PrettyPlots.allPlots — Method.

allPlots(n;idx)

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/10/2017, Last Modified: 5/29/2017

source

PrettyPlots.controlPlot — Method.

ctrp=controlPlot(n,idx,ctr); ctrp=controlPlot(n,idx,ctr,ctrp;(:append=>true));

to plot control signals

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/10/2017, Last Modified: 5/28/2017

source

PrettyPlots.currentSettings — Method.

currentSettings()

show the current plot settings

source

PrettyPlots.mainSim — Method.

mainSim(n,c,pa;(:mode=>:open1))

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 4/13/2017, Last Modified: 5/1/2017

source

PrettyPlots.obstaclePlot — Method.

to visualize the current obstacle in the field

obstaclePlot(n,c)

to run it after a single optimization

pp=obstaclePlot(n,c,1);

to create a new plot

pp=obstaclePlot(n,c,idx);

to add to an exsting position plot

pp=obstaclePlot(n,c,idx,pp;(:append=>true));

posterPlot

pp=obstaclePlot(n,c,ii,pp;(:append=>true),(:posterPlot=>true)); # add obstacles

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/11/2017, Last Modified: 4/3/2017

source

PrettyPlots.plotSettings — Method.

plotSettings(;(:simulate=>true)) plotSettings(;(:mpc_lines =>[(4.0,:blueviolet,:solid),(4.0,:blueviolet,:dash)]), (:plant_lines =>[(3.0,:darkgreen,:solid),(3.0,:darkgreen,:solid)]), (:limit_lines =>[(2.0,:turquoise,:solid),(2.0,:violet,:solid),(2.0,:orchid,:solid),(2.0,:darksalmon,:solid)]), (:size=>(1000,1000)) ) plotSettings(;(:obstacle_marker =>(:circle,:red,10.0,1.0)), (:goal_marker=>(:circle,:green,10.0,1.0)) ) plotSettings(;(:mpc_lines =>[(4.0,:blueviolet,:solid)]),(:size=>(700,700)))

this function can be called to modify the default settings in PrettyPlots.jl

check out Colors.jl for an incredible amount of colors!

source

PrettyPlots.posPlot — Method.

to plot the second solution

pp=posPlot(n,c,2) pp=posPlot(n,c,idx) –––––––––––––––––––––––––––––––––––––––––––

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/28/2017, Last Modified: 5/1/2017

source

PrettyPlots.posterP — Method.

default(guidefont = font(17), tickfont = font(15), legendfont = font(12), titlefont = font(20)) s=Settings(;save=true,MPC=true,simulate=false,format=:png,plantOnly=true); posterP(n,c,pa) –––––––––––––––––––––––––––––––––––––––––––

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 4/13/2017, Last Modified: 4/13/2017

source

PrettyPlots.statePlot — Method.

stp=statePlot(n,idx,st1,st2); stp=statePlot(n,idx,st1,st2;(:legend=>"test1")); stp=statePlot(n,idx,st1,st2,stp;(:append=>true),(:lims=>false));

to compare two different states

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/10/2017, Last Modified: 5/28/2017

source

PrettyPlots.statePlot — Method.

stp=statePlot(n,r.eval_num,7); stp=statePlot(n,idx,st); stp=statePlot(n,idx,st;(:legend=>"test1")); stp=statePlot(n,idx,st,stp;(:append=>true)); –––––––––––––––––––––––––––––––––––––––––––

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 2/10/2017, Last Modified: 5/28/2017

source

PrettyPlots.tPlot — Method.

tp=tPlot(n,r,idx) tp=tPlot(n,r,idx,tp;(:append=>true))

plot the optimization times

this is an MPC plot

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/11/2017, Last Modified: 3/11/2017

source

PrettyPlots.trackPlot — Method.

to visualize the current track in the field

trackPlot(c)

pp=trackPlot(c,pp;(:append=>true));

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/28/2017, Last Modified: 5/1/2017

source

PrettyPlots.vtPlot — Method.

vt=vtPlot(n,pa,c,idx)

Author: Huckleberry Febbo, Graduate Student, University of Michigan Date Create: 3/11/2017, Last Modified: 3/11/2017

source